feat: convert openai related tests to VCR
This converts OpenAI tests (including Ollama) to VCR instead
of manual patching. What this does is allow us to capture real
responses and then be able to know the difference in format
between the openai platform and others like Azure OpenAI or
clones like ollama or localai.

The first time a VCR test is run, it makes a call to the real
service, and then afterwards it re-uses the data. Since the
data remains, this also helps us safely refactor as otherwise
only those with accounts will know if changes break.

See https://pytest-vcr.readthedocs.io/en/latest/

Signed-off-by: Adrian Cole <adrian.cole@elastic.co>