fix: accidental loop on client-side validation fail
Before, if a tool use response failed validation (e.g. due to
being too large), we accidentally fed the error about the
validation back into the LLM. Most models would fail on that,
but for example llama3.1 instruct would ask a question and that
would end up recursing potentially 128 times (max_tool_use).

The intent I think was that when there was a response we feel
invalid that we would actually break the loop calling the LLM. So,
this breaks the loop in a way that retains the error and the stack
trace.

The reason this went unnoticed was two fold. First, as described
above, most LLMs would only recurse once before failing on a
ToolCallResult (due to being passed an error as input). Second,
we didn't check the type of response we received on such an error.
If we did, we would expect that a user-side validation error (so,
a pseudo-user reply), not an assistant error due to bad input.

Signed-off-by: Adrian Cole <adrian.cole@elastic.co>