test: refactor OpenAI tests to reduce redundancy
This refactors the OpenAI tests to reduce copy/paste redundancy.
Specifically, this changes to parameterized tests, with some
extra special-case handling for VCR. As a result, was are able
to have all the openai test deviations in the same file, but not
a whole lot of scanning to figure out what's different.

Signed-off-by: Adrian Cole <adrian.cole@elastic.co>